[+] #704 enforce AEAD integrity limit per RFC 9001 §6.6

+++ src/transport/xqc_conn.h
@@ -413,7 +413,7 @@ struct xqc_connection_s {
 
     /* for qlog */
     uint32_t                        MTU_updated_count;    
-    uint32_t                        packet_dropped_count;
+    uint64_t                        packet_dropped_count;  /**< RFC 9001 §6.6: decryption failure count across all keys */
     
     
     const
